+++ enterprise/cmd/frontend/internal/executor/init.go
@@ -9,11 +9,11 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/enterprise"
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/config"
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/queues/batches"
-	codeintel2 "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/queues/codeintel"
-	apiserver "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/server"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/codeintel"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/executor/config"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/executor/queues/batches"
+	codeintel2 "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/executor/queues/codeintel"
+	apiserver "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/executor/server"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbutil"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
 	"github.com/sourcegraph/sourcegraph/internal/oobmigration"
