+++ enterprise/cmd/frontend/internal/executor/internal_proxy_handler.go
@@ -14,7 +14,7 @@ import (
 	"github.com/inconshreveable/log15"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/envvar"
-	apiserver "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/executor-queue/server"
+	apiserver "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/executor/server"
 )
 
 func newInternalProxyHandler(uploadHandler http.Handler, queueOptions map[string]apiserver.QueueOptions) (func() http.Handler, error) {
